test: 통합 테스트 작성
