+++ src/main/resources/templates/admin/reservation.html
@@ -26,7 +26,7 @@
                 <a class="nav-link" href="/reservation">Reservation</a>
             </li>
             <li class="nav-item">
-                <a class="nav-link" href="/reservationTime">Time</a>
+                <a class="nav-link" href="/admin/time">Time</a>
             </li>
         </ul>
     </div>
